+++ src/resource/task.py
@@ -5,7 +5,7 @@ from flask_jwt_extended import jwt_required
 from flask_restful_swagger_3 import Resource, swagger, Api
 
 from src.model.task import Task
-from src.resource import clean_dict_input, add_swagger
+from src.resource import clean_dict_input, add_swagger, check_data_ownership
 from src.schema import ErrorSchema, SuccessSchema
 from src.swagger_patches import Schema, summary
 
@@ -83,14 +83,14 @@ class TaskResource(Resource):
 
         if 'building_id' in data:
             building_id = int(data.pop('building_id'))
-            from src.model.placeable.building import Building
-            building = current_app.db.session.query(Building).get(building_id)
+            from src.model.placeable.placeable import Placeable
+            building = current_app.db.session.query(Placeable).get(building_id)
             if building is None:
-                return ErrorSchema(message='Building id not found'), 400
+                return ErrorSchema(message='Placeable id not found'), 400
             if building.island_id != data['island_id']:
-                return ErrorSchema(message='Building does not belong to this island_id'), 400
-            if new_task and building.task is not None and building.task.is_running():
-                return ErrorSchema(message='Building is already being worked on'), 409
+                return ErrorSchema(message='Placeable does not belong to this island_id'), 400
+            if new_task and building.task is not None and not building.task.is_over():
+                return ErrorSchema(message='Placeable is already being worked on'), 409
 
             data['working_building'] = building
 
@@ -121,7 +121,10 @@ class TaskResource(Resource):
     @summary('Create a new task object')
     @swagger.response(response_code=200, description='Task object', schema=TaskSchema)
     @swagger.response(response_code=400, description='Unknown building id (when provided) or invalid task data', schema=ErrorSchema)
-    @swagger.response(response_code=409, description='Building is already being worked on', schema=ErrorSchema)
+    @swagger.response(response_code=409, description='Placeable is already being worked on', schema=ErrorSchema)
+    @swagger.response(response_code=403,
+                      description='Unauthorized access to data object. Calling user is not owner of the data (by island_id) (or admin)',
+                      schema=ErrorSchema)
     @swagger.expected(TaskSchema, required=True)
     @jwt_required()
     def post(self):
@@ -139,6 +142,11 @@ class TaskResource(Resource):
                 return r
 
             task = Task(**data)
+
+            r = check_data_ownership(
+                task.island_id)  # island_id == owner_id
+            if r: return r
+
             current_app.db.session.add(task)
             current_app.db.session.commit()
 
@@ -152,6 +160,10 @@ class TaskResource(Resource):
     @swagger.response(response_code=200, description='Task object', schema=TaskSchema)
     @swagger.response(response_code=400, description='Unknown building id (when provided) or invalid task data', schema=ErrorSchema)
     @swagger.response(response_code=404, description='Task not found', schema=ErrorSchema)
+    @swagger.response(response_code=409, description='Placeable is already being worked on', schema=ErrorSchema)
+    @swagger.response(response_code=403,
+                      description='Unauthorized access to data object. Calling user is not owner of the data (or admin)',
+                      schema=ErrorSchema)
     @swagger.expected(TaskSchema, required=True)
     @jwt_required()
     def put(self):
@@ -173,6 +185,10 @@ class TaskResource(Resource):
             if r is not None:
                 return r
 
+            r = check_data_ownership(
+                task.island_id)  # island_id == owner_id
+            if r: return r
+
             task.update(data)
             current_app.db.session.commit()
 
@@ -187,6 +203,9 @@ class TaskResource(Resource):
     @swagger.response(response_code=200, description='Task object', schema=SuccessSchema)
     @swagger.response(response_code=404, description='Task not found', schema=ErrorSchema)
     @swagger.response(response_code=400, description='No task id provided', schema=ErrorSchema)
+    @swagger.response(response_code=403,
+                      description='Unauthorized access to data object. Calling user is not owner of the data (or admin)',
+                      schema=ErrorSchema)
     @jwt_required()
     def delete(self):
         """
@@ -200,6 +219,10 @@ class TaskResource(Resource):
         task = Task.query.get(id)
         if task is None:
             return ErrorSchema(message='Task not found'), 404
+
+        r = check_data_ownership(
+            task.island_id)  # island_id == owner_id
+        if r: return r
 
         current_app.db.session.delete(task)
         current_app.db.session.commit()
